pyproject.toml
[tool.poetry]
name = "rai-latam-demo-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.0"
python-dotenv = "1.0.0"
PyYAML = "^6.0"
langchain = "^0.3.0"
langchain-groq = "^0.2.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
langsmith = "^0.4.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
